MBL-13944: Addressed PR feedback: Moved dataseeding APIs to test-driver area.
