+++ apps/flutter_parent/test_driver/apis/user_seed_api.dart
@@ -25,7 +25,7 @@ import 'package:flutter_parent/models/serializers.dart';
 import 'package:flutter_parent/network/utils/api_prefs.dart';
 import 'package:flutter_parent/network/utils/dio_config.dart';
 
-import '../auth_api.dart';
+import '../../lib/network/api/auth_api.dart';
 
 const _REDIRECT_URI = "urn:ietf:wg:oauth:2.0:oob";
